wip dialog
implement dialog and node delete in progress
